fix operator-sdk olm install
by making it use a proper HTTP client for building the
kubernetes client, ensuring OLM apis are added to the
scheme.Scheme used with the kubernetes client, and
splitting CRD and resource checks and installation
into distinct steps. This ensures CRDs are installed
and ready prior to the creation of associated CRs.

Signed-off-by: everettraven <everettraven@gmail.com>